refactor: move `getUpdatedCartItems` func to own file
