+++ imports/plugins/core/taxes/server/no-meteor/startup.js
@@ -1,50 +1,5 @@
 import { isEqual } from "lodash";
-import xformCartGroupToCommonOrder from "/imports/plugins/core/cart/server/no-meteor/util/xformCartGroupToCommonOrder";
-
-/**
- * @summary Returns `cart.items` with tax-related props updated on them
- * @param {Object} cart The cart
- * @param {Object} context App context
- * @returns {Object[]} Updated items array
- */
-async function getUpdatedCartItems(cart, context) {
-  const taxResultsByGroup = await Promise.all((cart.shipping || []).map(async (group) => {
-    const order = await xformCartGroupToCommonOrder(cart, group, context);
-    return context.mutations.getFulfillmentGroupTaxes(context, { order, forceZeroes: false });
-  }));
-
-  // Add tax properties to all items in the cart, if taxes were able to be calculated
-  const cartItems = (cart.items || []).map((item) => {
-    const newItem = { ...item };
-    taxResultsByGroup.forEach((group) => {
-      const matchingGroupTaxes = group.itemTaxes.find((groupItem) => groupItem.itemId === item._id);
-      if (matchingGroupTaxes) {
-        newItem.tax = matchingGroupTaxes.tax;
-        newItem.taxableAmount = matchingGroupTaxes.taxableAmount;
-        newItem.taxes = matchingGroupTaxes.taxes;
-      }
-    });
-    return newItem;
-  });
-
-  // Merge all group tax summaries to a single one for the whole cart
-  let combinedSummary = { tax: 0, taxableAmount: 0, taxes: [] };
-  for (const { taxSummary } of taxResultsByGroup) {
-    // groupSummary will be null if there wasn't enough info to calc taxes
-    if (!taxSummary) {
-      combinedSummary = null;
-      break;
-    }
-
-    combinedSummary.calculatedAt = taxSummary.calculatedAt;
-    combinedSummary.calculatedByTaxServiceName = taxSummary.calculatedByTaxServiceName;
-    combinedSummary.tax += taxSummary.tax;
-    combinedSummary.taxableAmount += taxSummary.taxableAmount;
-    combinedSummary.taxes = combinedSummary.taxes.concat(taxSummary.taxes);
-  }
-
-  return { cartItems, taxSummary: combinedSummary };
-}
+import getUpdatedCartItems from "./util/getUpdatedCartItems";
 
 const EMITTED_BY_NAME = "TAXES_CORE_PLUGIN";
 
@@ -64,7 +19,7 @@ export default function startup(context) {
   appEvents.on("afterCartUpdate", async ({ cart }, { emittedBy } = {}) => {
     if (emittedBy === EMITTED_BY_NAME) return; // short circuit infinite loops
 
-    const { cartItems, taxSummary } = await getUpdatedCartItems(cart, context);
+    const { cartItems, taxSummary } = await getUpdatedCartItems(context, cart);
 
     if (isEqual(cartItems, cart.items) && isEqual(taxSummary, cart.taxSummary)) return;
